requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32